Section 23 management
Added the management of the Section23

+++ PSE.Model/Output/Interfaces/ISection23.cs
@@ -28,7 +28,7 @@
 
         string? Sector { get; set; }
 
-        int? PercentShares { get; set; }
+        decimal? PercentShares { get; set; }
 
     }
